Fork the ESynth contract and use an allocator
For the ESynth and remove the methods for allocating funds in vaults.
Instead, use an allocator contract, which receives minted fund and
allocates them in vaults by the owner. The allocator is also responsible
for withdrawing the accumulated interest and putting it into a DSR
module after accounting for the fees.

This makes the ESynth contract simpler and allows to interate on the
allocation strategies.